requirements.txt
flask
invoke
asyncio==3.4.3
asyncpg==0.27.0
cloud-sql-python-connector==1.2.3
numpy==1.22.4
pandas==1.5.3
transformers==4.30.1
google-cloud-aiplatform
faker
langchain_google_vertexai
langchain==0.1.12
langchain_community
langchain_core
langchain_google_vertexai
pydantic>=2
langchain_astradb 
gunicorn
pyshorteners
urllib3<2.0